+++ LethalLevelLoader/Components/ExtendedContent/ExtendedLevel.cs
@@ -239,12 +239,6 @@ namespace LethalLevelLoader
                 }
             }
 
-            if (ContentType is ContentType.External)
-            {
-                OverrideDynamicRiskLevelAssignment = true;
-                UseTerrainFootsteps = false;
-            }
-
             SelectableLevel.spawnableMapObjects ??= [];
             SelectableLevel.indoorMapHazards ??= [];